feat(catalog-ui): New Digital Assistants Page
