+++ ui/catalog/src/constants/endpoints.constants.ts
@@ -1,7 +1,7 @@
 export const ROUTES = {
   LOGIN: "/login",
   LOGOUT: "/logout",
-  AI_DEPLOYMENTS: "/ai-deployments",
+  DIGITAL_ASSISTANTS: "/digital-assistants",
   ARCHITECTURES: "/architectures",
   SERVICES: "/services",
   USE_CASE_REFERENCES: "/use-case-references",
