--- ui/catalog/src/pages/AiDeployments/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./AiDeployments";
