+++ ui/catalog/src/pages/DigitalAssistants/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./DigitalAssistants";
